+++ ft_memcpy.c
@@ -6,15 +6,15 @@
 /*   By: toespino <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/14 18:35:08 by toespino          #+#    #+#             */
-/*   Updated: 2025/11/04 18:54:20 by toespino         ###   ########.fr       */
+/*   Updated: 2026/01/04 23:55:41 by toespino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memcpy(void *dest, const void *src, size_t n)
+void	*ft_memcpy(void *dest, const void *src, uint64_t n)
 {
-	size_t				i;
+	uint64_t				i;
 	unsigned char		*ptr;
 	const unsigned char	*ptr2;
