+++ ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: toespino <toespino@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/23 13:45:22 by toespino          #+#    #+#             */
-/*   Updated: 2025/10/24 17:57:44 by toespino         ###   ########.fr       */
+/*   Updated: 2026/01/05 00:05:31 by toespino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 char	*ft_strdup(const char *s)
 {
-	size_t	len;
-	size_t	i;
-	char	*out;
+	uint64_t	len;
+	uint64_t	i;
+	char		*out;
 
 	len = ft_strlen((char *)s);
 	out = malloc(len + 1);
