+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: toespino <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/14 11:20:39 by toespino          #+#    #+#             */
-/*   Updated: 2025/10/29 14:16:47 by toespino         ###   ########.fr       */
+/*   Updated: 2026/01/05 00:10:15 by toespino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *str)
+uint64_t	ft_strlen(const char *str)
 {
-	size_t	i;
+	uint64_t	i;
 
 	i = 0;
 	while (str[i])
