+++ ft_strncmp.c
@@ -6,15 +6,15 @@
 /*   By: toespino <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/16 08:36:40 by toespino          #+#    #+#             */
-/*   Updated: 2025/10/20 10:40:47 by toespino         ###   ########.fr       */
+/*   Updated: 2026/01/05 00:16:17 by toespino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strncmp(const char *s1, const char *s2, size_t n)
+int	ft_strncmp(const char *s1, const char *s2, uint64_t n)
 {
-	size_t	i;
+	uint64_t	i;
 
 	i = 0;
 	while (i < n && (s1[i] || s2[i]))
